// src/features/admin-partner/types/enum.ts
export enum PartnerStatus {
  ACTIVE = "อยู่ในความร่วมมือ",
  INACTIVE = "ไม่อยู่ในความร่วมมือ"
}

export enum PartnerType {
  SHOP = "ร้านค้า",
  AVENUE = "สถานที่จัดงาน",
  ALL = "ทั้งหมด"
}